to do: validators for new password and updating favorites section in real time
