+++ coctaildebucle.client/src/app/components/profile-update-form/profile-update-form.component.html
@@ -1,0 +1,53 @@
+  <!--  Update Profile -->
+<section class="profile-update">
+  <h3>{{ 'USER_PROFILE.UPDATE_INFORMATION' | translate }}</h3>
+
+  <form [formGroup]="profileForm" (ngSubmit)="onProfileUpdate()">
+    <div class="form-row" *ngFor="let field of editableFields">
+      <label class="form-label">{{ field.labelKey | translate }}</label>
+      <input [type]="field.type"
+             [formControlName]="field.controlName"
+             [readonly]="!field.editing"
+             class="form-control input-fixed" />
+      <button type="button" class="edit-btn" (click)="toggleEdit(field)">
+        {{ field.editing ? ('USER_PROFILE.CANCEL_BUTTON' | translate) : ('USER_PROFILE.EDIT_BUTTON' | translate) }}
+      </button>
+    </div>
+
+
+    <!-- gdpr field -->
+    <div class="form-row">
+      <label class="form-label">{{ 'USER_PROFILE.FIELD_GDPR' | translate }}</label>
+      <input type="checkbox" formControlName="gdprConsent" />
+    </div>
+
+    <!-- password field -->
+    <div class="form-row">
+      <label class="form-label">{{ 'USER_PROFILE.FIELD_PASS_CURRENT' | translate }}</label>
+      <input type="password"
+             value="********"
+             readonly
+             class="form-control input-fixed" />
+      <button type="button" class="edit-btn" (click)="togglePasswordEdit()">
+        {{ passwordEditMode ? ('USER_PROFILE.CANCEL_BUTTON' | translate) : ('USER_PROFILE.EDIT_BUTTON' | translate) }}
+      </button>
+    </div>
+
+    <div *ngIf="passwordEditMode">
+      <div class="form-row">
+        <label class="form-label">{{ 'USER_PROFILE.FIELD_PASS_NEW' | translate }}</label>
+        <input type="password" id="newPassword" formControlName="newPassword" class="form-control input-fixed" />
+      </div>
+
+      <div class="form-row">
+        <label class="form-label">{{ 'USER_PROFILE.FIELD_PASS_CONFIRM' | translate }}</label>
+        <input type="password" id="confirmNewPassword" formControlName="confirmNewPassword" class="form-control input-fixed" />
+        <div *ngIf="passwordMismatch" class="text-danger">{{ 'USER_PROFILE.FIELD_PASS_MISMATCH' | translate }}</div>
+      </div>
+    </div>
+
+    <button type="submit" class="btn btn-primary save-btn" [disabled]="profileForm.invalid">{{ 'USER_PROFILE.SAVE_INFORMATION' | translate }}</button>
+    <div *ngIf="successMessage" class="text-success">{{ successMessage }}</div>
+    <div *ngIf="errorMessage" class="text-danger">{{ errorMessage }}</div>
+  </form>
+</section>
