+++ coctaildebucle.client/src/app/components/user-profile/user-profile.component.css
@@ -61,60 +61,6 @@
   transform: scale(1.05);
 }
 
-.profile-update
-{
-  margin: 2rem auto;
-  max-width: 600px;
-  background-color: #f9f9f9;
-  padding: 2rem;
-  border-radius: 10px;
-  box-shadow: 0px 4px 10px rgba(0,0,0,0.1);
-}
-
-.form-row
-{
-  display: flex;
-  align-items: center;
-  margin-bottom: 1.5rem;
-}
-
-.form-label
-{
-  flex: 0 0 150px;
-  text-align: right;
-  margin-right: 20px;
-  font-weight: bold;
-}
-
-.input-fixed {
-  flex: 1;
-  padding: 0.5rem;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-  min-width: 250px;
-}
-
-.edit-btn {
-  margin-left: 1rem;
-  padding: 0.5rem 1rem;
-  background-color: #007bff;
-  color: white;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-}
-
-.edit-btn:hover {
-  background-color: #0056b3;
-}
-
-.save-btn {
-  display: block;
-  margin: 2rem auto 0 auto;
-  padding: 0.75rem 2rem;
-  font-size: 1.2rem;
-}
-
 /* Created Drinks Section */
 
 .drink-container {
@@ -146,9 +92,4 @@ img {
   padding: 0.5em !important;
   border-radius: 1.5em !important;
   /*   margin-bottom: 10px;*/
-}
-
-.backBtn
-{
-    margin-top:5em;
 }
